+++ src/views/mod.rs
@@ -1,2 +1,3 @@
 pub mod app;
 pub mod the_canvas;
+pub mod the_toolbar;
